+++ examples/nova_carter_demo/scripts/run_l4t_docker.sh
@@ -9,6 +9,5 @@ docker run \
     --network host \
     --runtime nvidia \
     -v $(pwd):/remembr \
-    --device=/dev/video0:/dev/video0 \
     -w /remembr/examples/nova_carter_demo \
     nova_carter_demo:l4t