+++ remembr/agents/remembr_agent.py
@@ -339,6 +339,7 @@ class ReMEmbRAgent(Agent):
 
         # let us parse and check the output is a dictionary. raise error otherwise
         response = ''.join(response.content.splitlines())
+        print("response: ", response)
 
         try:
             if '```json' not in response:
@@ -354,8 +355,19 @@ class ReMEmbRAgent(Agent):
                 if key not in parsed:
                     raise ValueError("Missing all the required keys during generate. Retrying...")
         except:
-            raise ValueError("Generate call failed. Retrying...")
-                
+            parsed = {
+                'type_reasoning': '',
+                'type': '',
+                'answer_reasoning': '',
+                'text': response,
+                'keypoint': '',
+                'binary': None,
+                'position': None,
+                'orientation': None,
+                'time': None,
+                'duration': None}
+            # raise ValueError("Generate call failed. Retrying...")
+
         try:
             if type(parsed['position']) == str:
                 parsed['position'] = eval(parsed['position'])
@@ -371,6 +383,28 @@ class ReMEmbRAgent(Agent):
             print("Position parsing failled", parsed['position'])
             parsed['position'] = [None, None, None]
 
+        try:
+            if parsed['keypoint']:
+                for message in messages:
+                    if parsed['keypoint'] in message.content:
+                        print("fullfill by ", message)
+                        parsed['keypoint'] = message.content
+                        break
+            if None in parsed['position'] and parsed['keypoint']:
+                match = re.search(r"\[(-?\d+\.\d+,\s*-?\d+\.\d+,\s*-?\d+\.\d+)\]", parsed['keypoint'])
+                print("re position: ", match)
+                if match:
+                    position = match.group(1)  # Extract matched position
+                    position = [float(num) for num in position.split(", ")]  # Convert to a list of floats
+                    parsed['position'] = position
+            
+
+
+            
+        
+        except:
+            pass
+                
         self.previous_tool_requests = "These are the tools I have previously used so far: \n"
         self.agent_call_count = 0